Call expressions must be allowed as top-level statements in functions
Currently, the following fails to parse:
void foo() {
    bar();
}

This is because we limit top-level statements in functions to "effectful" statements.
We do this to eliminate ambiguity of the form "a * b". This could either be a declaration
of a pointer named "b" or describing the computation of "multiply a times b". However,
because we have pointers, a function call is able to assign to a variable in the current
scope, so it needs to be considered an "effectful" statement.