Replace the control-flow stack by a divergence stack
The problem this is trying to fix is defining uniform control-flow for
the purpose of giving the semantics of control barriers and derivatives.

I originally thought that merely considering the values we diverge on
would be enough, but the following example show that we must also
remember where we diverged:
void foo() {
    control_barrier();
}
void bar(int tid) {
    int i = 0;
    do {
        if ((tid + i) % 2 == 0)
            if (true)
                foo();
        if ((tid + i) % 2 == 1)
            foo();
        i++;
    } while (i < 2);
}

In this patch, I fixed the spec, but not yet the ott.